refactor: custom hook for create step and card
