+++ apps/journeys-admin/src/components/Editor/Slider/JourneyFlow/JourneyFlow.tsx
@@ -24,28 +24,22 @@ import {
   useEdgesState,
   useNodesState
 } from 'reactflow'
-import { v4 as uuidv4 } from 'uuid'
 
 import { useEditor } from '@core/journeys/ui/EditorProvider'
 import { useJourney } from '@core/journeys/ui/JourneyProvider'
-import { searchBlocks } from '@core/journeys/ui/searchBlocks'
 import ArrowRefresh6Icon from '@core/shared/ui/icons/ArrowRefresh6'
 
 import {
   GetStepBlocksWithPosition,
   GetStepBlocksWithPositionVariables
 } from '../../../../../__generated__/GetStepBlocksWithPosition'
-import { ThemeMode, ThemeName } from '../../../../../__generated__/globalTypes'
-import { useBlockOrderUpdateMutation } from '../../../../libs/useBlockOrderUpdateMutation'
-import { useNavigateToBlockActionUpdateMutation } from '../../../../libs/useNavigateToBlockActionUpdateMutation'
-import { useStepAndCardBlockCreateMutation } from '../../../../libs/useStepAndCardBlockCreateMutation'
-import { useStepBlockNextBlockUpdateMutation } from '../../../../libs/useStepBlockNextBlockUpdateMutation'
 import { useStepBlockPositionUpdateMutation } from '../../../../libs/useStepBlockPositionUpdateMutation'
 
 import { CustomEdge } from './edges/CustomEdge'
 import { StartEdge } from './edges/StartEdge'
 import { PositionMap, arrangeSteps } from './libs/arrangeSteps'
 import { transformSteps } from './libs/transformSteps'
+import { useCreateStepAndCard } from './libs/useCreateStepAndCard'
 import { NewStepButton } from './NewStepButton'
 import { SocialPreviewNode } from './nodes/SocialPreviewNode'
 import { StepBlockNode } from './nodes/StepBlockNode'
@@ -68,8 +62,7 @@ export const GET_STEP_BLOCKS_WITH_POSITION = gql`
 export function JourneyFlow(): ReactElement {
   const { journey } = useJourney()
   const {
-    state: { steps },
-    dispatch
+    state: { steps }
   } = useEditor()
   const connectingParams = useRef<OnConnectStartParams | null>(null)
   const [nodes, setNodes, onNodesChange] = useNodesState([])
@@ -78,31 +71,8 @@ export function JourneyFlow(): ReactElement {
 
   const [reactFlowInstance, setReactFlowInstance] =
     useState<ReactFlowInstance | null>(null)
-  const [stepAndCardBlockCreate] = useStepAndCardBlockCreateMutation({
-    update(cache, { data }) {
-      if (data?.stepBlockCreate != null && data?.cardBlockCreate != null) {
-        cache.modify({
-          fields: {
-            blocks(existingBlockRefs = []) {
-              const newStepBlockRef = cache.writeFragment({
-                data: data.stepBlockCreate,
-                fragment: gql`
-                  fragment NewBlock on Block {
-                    id
-                  }
-                `
-              })
-              return [...existingBlockRefs, newStepBlockRef]
-            }
-          }
-        })
-      }
-    }
-  })
-  const [stepBlockNextBlockUpdate] = useStepBlockNextBlockUpdateMutation()
   const [stepBlockPositionUpdate] = useStepBlockPositionUpdateMutation()
-  const [navigateToBlockActionUpdate] = useNavigateToBlockActionUpdateMutation()
-  const [blockOrderUpdate] = useBlockOrderUpdateMutation()
+  const createStepAndCard = useCreateStepAndCard()
 
   async function blockPositionsUpdate(positions: PositionMap): Promise<void> {
     if (steps == null || journey == null) return
@@ -191,100 +161,100 @@ export function JourneyFlow(): ReactElement {
     setNodes(nodes)
   }, [steps, data, setEdges, setNodes, refetch])
 
-  const createStepAndCardBlock = useCallback(
-    async function createStepAndCardBlock(
-      nodeId: string,
-      handleId: string,
-      x: number,
-      y: number
-    ): Promise<void> {
-      const step = steps?.find((step) => step.id === nodeId)
-      const block = searchBlocks(step != null ? [step] : [], handleId)
+  // const createStepAndCardBlock = useCallback(
+  //   async function createStepAndCardBlock(
+  //     nodeId: string,
+  //     handleId: string,
+  //     x: number,
+  //     y: number
+  //   ): Promise<void> {
+  //     const step = steps?.find((step) => step.id === nodeId)
+  //     const block = searchBlocks(step != null ? [step] : [], handleId)
 
-      if (nodeId !== 'SocialPreview' && (step == null || block == null)) return
-      if (journey == null) return
-      const newStepId = uuidv4()
-      const newCardId = uuidv4()
+  //     if (nodeId !== 'SocialPreview' && (step == null || block == null)) return
+  //     if (journey == null) return
+  //     const newStepId = uuidv4()
+  //     const newCardId = uuidv4()
 
-      const { data } = await stepAndCardBlockCreate({
-        variables: {
-          stepBlockCreateInput: {
-            id: newStepId,
-            journeyId: journey.id,
-            x,
-            y
-          },
-          cardBlockCreateInput: {
-            id: newCardId,
-            journeyId: journey.id,
-            parentBlockId: newStepId,
-            themeMode: ThemeMode.dark,
-            themeName: ThemeName.base
-          }
-        }
-      })
+  //     const { data } = await stepAndCardBlockCreate({
+  //       variables: {
+  //         stepBlockCreateInput: {
+  //           id: newStepId,
+  //           journeyId: journey.id,
+  //           x,
+  //           y
+  //         },
+  //         cardBlockCreateInput: {
+  //           id: newCardId,
+  //           journeyId: journey.id,
+  //           parentBlockId: newStepId,
+  //           themeMode: ThemeMode.dark,
+  //           themeName: ThemeName.base
+  //         }
+  //       }
+  //     })
 
-      if (data != null) {
-        dispatch({
-          type: 'SetSelectedStepAction',
-          selectedStep: {
-            ...data.stepBlockCreate,
-            children: [{ ...data.cardBlockCreate, children: [] }]
-          }
-        })
-      }
+  //     if (data != null) {
+  //       dispatch({
+  //         type: 'SetSelectedStepAction',
+  //         selectedStep: {
+  //           ...data.stepBlockCreate,
+  //           children: [{ ...data.cardBlockCreate, children: [] }]
+  //         }
+  //       })
+  //     }
 
-      if (nodeId === 'SocialPreview') {
-        // social preview
-        await blockOrderUpdate({
-          variables: {
-            id: newStepId,
-            journeyId: journey.id,
-            parentOrder: 0
-          },
-          optimisticResponse: {
-            blockOrderUpdate: [
-              {
-                id: newStepId,
-                __typename: 'StepBlock',
-                parentOrder: 0
-              }
-            ]
-          }
-        })
-      } else if (step != null && step.id === block?.id) {
-        // step
-        await stepBlockNextBlockUpdate({
-          variables: {
-            id: step.id,
-            journeyId: journey.id,
-            input: {
-              nextBlockId: newStepId
-            }
-          },
-          optimisticResponse: {
-            stepBlockUpdate: {
-              id: step.id,
-              __typename: 'StepBlock',
-              nextBlockId: newStepId
-            }
-          }
-        })
-      } else if (block != null) {
-        // action
-        await navigateToBlockActionUpdate(block, newStepId)
-      }
-    },
-    [
-      journey,
-      steps,
-      navigateToBlockActionUpdate,
-      stepAndCardBlockCreate,
-      stepBlockNextBlockUpdate,
-      blockOrderUpdate,
-      dispatch
-    ]
-  )
+  //     if (nodeId === 'SocialPreview') {
+  //       // social preview
+  //       await blockOrderUpdate({
+  //         variables: {
+  //           id: newStepId,
+  //           journeyId: journey.id,
+  //           parentOrder: 0
+  //         },
+  //         optimisticResponse: {
+  //           blockOrderUpdate: [
+  //             {
+  //               id: newStepId,
+  //               __typename: 'StepBlock',
+  //               parentOrder: 0
+  //             }
+  //           ]
+  //         }
+  //       })
+  //     } else if (step != null && step.id === block?.id) {
+  //       // step
+  //       await stepBlockNextBlockUpdate({
+  //         variables: {
+  //           id: step.id,
+  //           journeyId: journey.id,
+  //           input: {
+  //             nextBlockId: newStepId
+  //           }
+  //         },
+  //         optimisticResponse: {
+  //           stepBlockUpdate: {
+  //             id: step.id,
+  //             __typename: 'StepBlock',
+  //             nextBlockId: newStepId
+  //           }
+  //         }
+  //       })
+  //     } else if (block != null) {
+  //       // action
+  //       await navigateToBlockActionUpdate(block, newStepId)
+  //     }
+  //   },
+  //   [
+  //     journey,
+  //     steps,
+  //     navigateToBlockActionUpdate,
+  //     stepAndCardBlockCreate,
+  //     stepBlockNextBlockUpdate,
+  //     blockOrderUpdate,
+  //     dispatch
+  //   ]
+  // )
   const onConnect = useCallback<OnConnect>(() => {
     // reset the start node on connections
     connectingParams.current = null
@@ -325,7 +295,7 @@ export function JourneyFlow(): ReactElement {
           y: (event as unknown as MouseEvent).clientY
         })
 
-        void createStepAndCardBlock(
+        void createStepAndCard(
           connectingParams.current.nodeId,
           connectingParams.current.handleId,
           parseInt(x.toString()),
@@ -333,7 +303,7 @@ export function JourneyFlow(): ReactElement {
         )
       }
     },
-    [reactFlowInstance, connectingParams, createStepAndCardBlock]
+    [reactFlowInstance, connectingParams, createStepAndCard]
   )
   const onNodeDragStop: NodeDragHandler = async (
     _event,
